+++ code/payment-service/src/main/java/vn/id/thongdanghoang/sep/payment_service/PromotionApplied.java
@@ -1,10 +1,12 @@
 package vn.id.thongdanghoang.sep.payment_service;
 
+import java.math.BigDecimal;
+
 public record PromotionApplied(
         String transactionId,
         String userId,
-        double originalAmount,
-        double discountAmount,
-        double finalAmount,
+        BigDecimal originalAmount,
+        BigDecimal discountAmount,
+        BigDecimal finalAmount,
         String appliedVoucher
 ) {}
